% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gcr.R
\name{gcr}
\alias{gcr}
\title{Generalized Correlation Regression}
\usage{
gcr(
  Y,
  X,
  W,
  alpha_init,
  beta_init,
  phi_init,
  phi.include = TRUE,
  family,
  accelerate = FALSE,
  lambda = 1,
  max_iter_1 = 100,
  max_iter_2 = 100,
  tol = 1e-06,
  eps = .Machine$double.eps^0.5,
  independent = FALSE,
  get.se = TRUE,
  criteria = "max",
  verbose = FALSE
)
}
\arguments{
\item{Y}{Response variable as a list of length \code{n}, where each element is a numeric vector
of responses for a cluster. For cluster \eqn{i}, \code{Y[[i]]} should be a vector of length \eqn{m_i}
(number of observations in cluster \eqn{i}).}

\item{X}{A list of length \code{n} containing moment-related covariate matrices for each cluster. Each element
\code{X[[i]]} should be a numeric matrix of dimension \eqn{m_i \times p} where:
\itemize{
  \item \eqn{m_i} is the number of observations in cluster \eqn{i}
  \item \eqn{p} is the number of covariates
}}

\item{W}{A list of length \code{n} containing correlation-related covariate matrices for each cluster. Each element
\code{W[[i]]} should be a numeric matrix of dimension \eqn{m_i(1-m_i)/2 \times d} where:
\itemize{
  \item \eqn{m_i} is the number of observations in cluster \eqn{i}
  \item \eqn{d} is the number of covariates
}}

\item{alpha_init}{Initial value for alpha}

\item{beta_init}{Initial value for beta}

\item{phi_init}{Initial value for phi}

\item{phi.include}{Logical indicating whether phi is to be estimated}

\item{family}{Response distribution family: "Gaussian", "Poisson", or "Binomial"}

\item{accelerate}{Logical indicating whether use the accelerated algorithm}

\item{lambda}{Step size}

\item{max_iter_1}{Maximum iterations for outer optimization loop}

\item{max_iter_2}{Maximum iterations for inner optimization loop}

\item{tol}{Convergence tolerance threshold}

\item{eps}{Difference step size in the Hessian approximation}

\item{independent}{Logical indicating whether use independent correlation structure}

\item{get.se}{Logical indicating whether covariance matrices and p-values}

\item{criteria}{Convergence criteria: "sum" (sum absolute changes),
"avg" (mean absolute change) or "max" (max absolute change)}

\item{verbose}{Logical indicating whether print convergence information}
}
\value{
A list containing the following components:
\describe{
  \item{alpha}{Estimated alpha}
  \item{pvalue_alpha}{p-value for estimated alpha}
  \item{beta}{Estimated beta}
  \item{pvalue_beta}{p-value for estimated beta}
  \item{phi}{Estimated phi}
  \item{H_1}{Estimated H_1}
  \item{H_mid}{Estimated H_mid}
  \item{H_2}{Estimated H_2}
  \item{Hessian}{Estimated Hessian matrix}
}
}
\description{
Fits a Generalized Correlation Regression model,
supporting multiple distribution families and convergence criteria.
}
